backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-admin-interface==0.17.1
django-colorfield==0.4.3
django-cors-headers==3.9.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==4.0.1
MarkupSafe==2.0.1
packaging==21.0
pycodestyle==2.7.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rich==10.14.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.7